stop stacking coordinate systems
I refactored the whole bounding_box and region thing to reduce the amount of math the code has to do.  now all bounding boxen are regarded to be in absolute map coordinates, and bounding box is considered to be an inheritable trait rather than something that stacks.  most importantly, this allows the full map to have empty space, which allows me to have a nice little border around the outside.  it also means there's no top-level group anymore, which is good because I thaut it was a waste of space.