reduce spacial precision
since the viewbox units are now all in millimeters, three digits after the decimal is a ridiculus amount of precision.  going down to the hundredth of a millimeter is still probably unnecessary, but since the line widths go down to of order a tenth of a millimeter our precision should probably be finer than that.