polish labels
the main code change here was making the units more consistent so I could use units when I set the font sizes.  you may note if you built this map that the labels are still not actually all placed.  I wanted to do it programaticly so that I can easily update and resize the map, but it was going to take *so* long.  maybe someday I'll make a script to let me graphically edit the config file, but for now I'm just going to do it in Illustrator.