fix units
I finally figured out how units work in SVG.  the unit names you put on "absolute lengths" are implemented as flat multipliers that depend on the resolution of your screen.  these multipliers are correct when SVG is in the default coordinate system, but applying a viewbox scales all lengths equally, regardless of whether they're unitless or "absolute".  thus, in an svg with a viewbox 1mm is typically a completely arbitrary length.  anyway, I set it to be as self-consistent as it can be.  now everything is in millimeters agen.  font sizes are required to have units (EVEN THO THE UNITS ARE NOT CORRECT IF YOU USE A VIEWBOX) so I put "px" on them.  don't be fooled; px is equivalent to millimeters (but mm would actually be 3mm so you can't use mm).