separate code into three distinct steps
I went to add graticules, and I realized it was going to be a pain because I would have to rewrite a bunch of the code relating to Layers and Groups for it.  so I did a big refactor.  now, instead of it loading, scaling, and plotting each shape one at a time, it first evaluates all Layers and Graticules into Groups, and then applies the coordinate transform, and then plots them all as SVG strings.  breaking it up like this has no effect on behavior besides forcing me to get rid of the nice print statements, but it makes the code nicer and easier to read, and will also make a great many future features easier to implement.