fix multiline markers
I needed to get markers with multiple lines of SVG to work so that I could use the MBTA logo as a marker.  there turned out to be a surprising number of issues with it.  but it works now.  and the indentation is fixed. (apparently the indentation was wrong before)