ensure labels are always on-screen
when you label polygons, the label can often end up off screen even if the polygon itself has some points on screen.  thats' not great becuase those labels then won't show up when you view the SVG, so you won't know to move them, and you'll end up with visible polygons that aren't labelled as well as labels that take up space in the file but don't contribute to the image.  anyway, I changed it so it doesn't do that anymore.