rotating markers
I implemented compass rose functionality.  what you do is add a marker with "bearing" set to 0.  then when it sees that the marker's orientation matters, it'll automaticly rotate it according to the Transform's jacobian matrix so it points in the desired direction.  it works great!  well, except that it doesn't work.  it's currently rotating the pointer just *slightly* too much, and I can't figure out why.  but it's close enuff for now so I'm pushing it.  I think I just need more tests for this jacobian function.