increase scale of Northeast map
this was tricky, because the Natural Earth dataset only goes up to 1:10M, whereas I wanted more like 1:1M, so the data was too coarse.  it turns out, tho, that the USGS publishes all the data I need at a nominal scale of 1:1M.  the bad news is that their rivers dataset has way too many rivers on it; to the extent that it's pretty much unusable.  luckily I found another dataset published by the Commission for Environmental Cooperation that has rivers more as I want them, but the issue that actually makes this all difficult is that their shapefiles are projected into a North American Lambert azimuthal projection.

so anyway, I spent a really long time studying different geodetic coordinate systems and writing code that would parse the shapefile's .prj file and automaticly remove the projection and convert from whatever datum it was using to WGS84, but then it was hard and I found out QGIS is free and can do that part for me.  so I used QGIS to manually convert the coordinate systems to WGS84, and now it looks great.  thanks, QGIS!