coerce labels into bounding_box, not region
for maps that used the datasets' bilt-in projections, the old system was fine, but I'm now having issues on my northeast map where labels get coerced into the graticule-alined region, but still aren't visible in the final map.  this fixes that.  there's also a 2mm buffer now to ensure that all autogenerated labels are actually fully visible, even if they're on the top or right side of the map.