a method of automatic city placement
I wrote a script to automatically choose cities from a large shapefile with many cities in it to give me a reasonably well distributed but fully automated distribution of labelled cities.  this was necessary because the natural earth shapefile I had been using was much too sparse.  I'm not sure if I'm going to keep it like this; I'm starting to think a more human touch is required to get a truly optimal distribution of labels, accounting for cultural ties, airports and amtrak stations, and whatnot, or at the very least a raster dataset of population density, but this is a great start, and I'm pretty happy with it so far.