refined taste in cities
I made the city selection algorithm better, so it shows more cities in general now but also does a better job of not showing cities that are really just the same city.  it now also ensures that annapolis is always shown.  it's not very big but I think it should be on the map since it is a state capital.