rotation!
I also redefined scale fo Affine transformations to account for an implicit unit conversion from meters to millimeters.  idk if that's a good idea, since you can't really use Affine for equirectangular projections anymore.  ideally I would be able to write the true scale in the yaml file and then have it automaticly change the scale when it loads it in, but I need to separate my deserialization from my data model for that.