choose streets more carefully
I did what I could to make the streets apprea more or less exactly how I want them in the camberville map.  there are still a few that I'll have to delete when I do the final map, but I think a manageable number.  unfortunately it now takes an extremely long time to generate the map, because I'm now loading the full road shapefile four times for reasons.  it's fine.  it's probably fine.