advanced rotation
I wrote a new script to help me iterate on the aspect for the oblique mercator projection for the northeast map.  using that  Icame up with a slitely less rotated version (that is, more north-up) that allows me to obscure both vermont and west virginia, while simultaneusly giving the Washington metro area more breathing room.