more illustrator compatibility
it turns out illustrator will ignore the alpha term in the color specifier, and also fails to reliably load paths above a certain number of vertices.

on the code side, it turns out illustrator doesn't support clip-paths with multiple elements, so if you try to create a clip-path with a group as its child, it now consolidates that group into a single path.

I also changed the railroad styling in the northeast map and lightened the street labels in the camberville and newton maps.