finish this map*
it's done!  or rather, it's as close as I care to get it to done because I'm tired of working on it now.  it's pretty close; I'm just missing the water in New Jersey, some of the bridges over the hudson river, and everything except the trains in Tokyo.  and a tiny bit of la provincia de buenos aires.  but overall, I'm quite happy with it.