some new maps
you'll note that I've added a new type of layer called "Hillshade" to the code.  I haven't implemented this yet because I haven't figured out how to read geotiff files, but I do intend to do this at some point.  for now, trying to add a Hillshade layer will throw a not-implemented error.